requirements.txt
pytest~=8.2.0
selenium~=4.20.0
allure-pytest~=2.13.5
